+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 
 # Change Log
 All notable changes to this project will be documented in this file.
+
+## [2.0.1] - 2026-02-09
+
+### Added
+- Last RSSI signal strength
+- Last sensors calibration date
+
+### Changed
+- Documentation
+- Changelog file
  
 ### [2.0.0]